apps/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.1
python-multipart>=0.0.9
pydantic>=2.10.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
